fixed-spec-error

+++ app/Http/Controllers/ProductSpecificationController.php
@@ -9,6 +9,10 @@ class ProductSpecificationController extends Controller
 {
     public function delete($id){
         $product = ProductSpecification::where('id',$id)->first();
+
+        if(!$product){
+            return redirect()->back()->with('message',['icon' => 'success', 'text' => 'Spec Not found']);
+        }
         $product->delete();
         return redirect()->back()->with('message',['icon' => 'success', 'text' => ' successfully deleted!']);
